Reorganize jpdf and statis_filt
Reorganize class methods in jpdf.py into functions because they do not logically constitute a class. And functions are more concise and straightforward.

Fix some bugs in statis_filt.py that are not campatible with the filtdns.py module. Use tools.py and fileIO.py modules to replace some functional static method in the class.

Add full wave numbers as a member of the basic dataset parameters.